feat: agregar soporte para roles de usuario en la consulta de cursos por tipo de programa
- Se actualizó la lógica en CourseController para manejar tanto estudiantes como profesores al obtener cursos.
- Se modificó la interfaz ICourseRepository para incluir el rol del usuario en la consulta de cursos.
- Se implementó la clase FindCoursesByProgramTypeQuery para gestionar la lógica de consulta según el rol.
- Se creó la clase GetCourseDetailQuery para obtener detalles del curso, incluyendo información del profesor responsable.
- Se refactorizó TypeormCoursesRepository para utilizar las nuevas clases de consulta.
- Se mejoró CourseService para aceptar el rol del usuario en la búsqueda de cursos.
- Se añadió CourseUtils para formatear fechas y horas, y obtener nombres de días en español.
